refactor(types): refactor `rmk-types` (#753)
