+++ rmk-types/src/lib.rs
@@ -25,8 +25,9 @@
 #![no_std]
 
 pub mod action;
+pub mod battery;
+pub mod ble;
 pub mod connection;
-pub mod event;
 pub mod fork;
 pub mod keycode;
 pub mod led_indicator;
