+++ rmk-types/src/protocol/rmk/config.rs
@@ -1,0 +1,79 @@
+//! Behavior configuration protocol types.
+
+use heapless::Vec;
+use serde::{Deserialize, Serialize};
+
+use super::{MAX_COMBO_KEYS, MAX_MORSE_PATTERNS};
+use crate::action::{KeyAction, MorseProfile};
+use crate::fork::ForkStateBits;
+use crate::modifier::ModifierCombination;
+
+/// Protocol-facing morse/tap-dance configuration.
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, postcard_schema::Schema)]
+pub struct MorseConfig {
+    pub profile: MorseProfile,
+    pub patterns: Vec<MorsePatternEntry, MAX_MORSE_PATTERNS>,
+}
+
+/// A single morse pattern/action pair.
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize, postcard_schema::Schema)]
+pub struct MorsePatternEntry {
+    pub pattern: u16,
+    pub action: KeyAction,
+}
+
+/// Protocol-facing fork (key override) configuration.
+///
+/// This mirrors firmware `Fork` fields without reducing match-state dimensions.
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize, postcard_schema::Schema)]
+pub struct ForkConfig {
+    pub trigger: KeyAction,
+    pub negative_output: KeyAction,
+    pub positive_output: KeyAction,
+    pub match_any: ForkStateBits,
+    pub match_none: ForkStateBits,
+    pub kept_modifiers: ModifierCombination,
+    pub bindable: bool,
+}
+
+/// Protocol-facing combo configuration.
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, postcard_schema::Schema)]
+pub struct ComboConfig {
+    pub actions: Vec<KeyAction, MAX_COMBO_KEYS>,
+    pub output: KeyAction,
+    pub layer: Option<u8>,
+}
+
+/// Protocol-facing behavior configuration.
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize, postcard_schema::Schema)]
+pub struct BehaviorConfig {
+    pub combo_timeout_ms: u16,
+    pub oneshot_timeout_ms: u16,
+    pub tap_interval_ms: u16,
+    pub tap_tolerance: u8,
+}
+
+/// Summary information about macro capabilities.
+///
+/// These fields are also available via [`DeviceCapabilities`]; this type
+/// provides a lightweight query for tools that only need macro info.
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize, postcard_schema::Schema)]
+pub struct MacroInfo {
+    pub max_macros: u8,
+    pub macro_space_size: u16,
+}
+
+impl From<super::DeviceCapabilities> for MacroInfo {
+    fn from(caps: super::DeviceCapabilities) -> Self {
+        Self {
+            max_macros: caps.max_macros,
+            macro_space_size: caps.macro_space_size,
+        }
+    }
+}
+
+/// Raw macro data for a single macro.
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, postcard_schema::Schema)]
+pub struct MacroData {
+    pub data: heapless::Vec<u8, { super::MAX_MACRO_DATA }>,
+}
