+++ rmk/src/event/state.rs
@@ -2,60 +2,59 @@
 
 use rmk_macro::event;
 use rmk_types::led_indicator::LedIndicator;
-use rmk_types::event::{LayerChangePayload, LedPayload, SleepPayload, WpmPayload};
 
 /// Active layer changed event
 #[event(channel_size = crate::LAYER_CHANGE_EVENT_CHANNEL_SIZE, pubs = crate::LAYER_CHANGE_EVENT_PUB_SIZE, subs = crate::LAYER_CHANGE_EVENT_SUB_SIZE)]
 #[derive(Clone, Copy, Debug, PartialEq, Eq)]
 #[cfg_attr(feature = "defmt", derive(defmt::Format))]
-pub struct LayerChangeEvent(pub LayerChangePayload);
+pub struct LayerChangeEvent(pub u8);
 
 impl LayerChangeEvent {
     pub fn new(layer: u8) -> Self {
-        Self(LayerChangePayload { layer })
+        Self(layer)
     }
 }
 
-impl_payload_wrapper!(LayerChangeEvent, LayerChangePayload);
+impl_payload_wrapper!(LayerChangeEvent, u8);
 
 /// WPM updated event
 #[event(channel_size = crate::WPM_UPDATE_EVENT_CHANNEL_SIZE, pubs = crate::WPM_UPDATE_EVENT_PUB_SIZE, subs = crate::WPM_UPDATE_EVENT_SUB_SIZE)]
 #[derive(Clone, Copy, Debug, PartialEq, Eq)]
 #[cfg_attr(feature = "defmt", derive(defmt::Format))]
-pub struct WpmUpdateEvent(pub WpmPayload);
+pub struct WpmUpdateEvent(pub u16);
 
 impl WpmUpdateEvent {
     pub fn new(wpm: u16) -> Self {
-        Self(WpmPayload { wpm })
+        Self(wpm)
     }
 }
 
-impl_payload_wrapper!(WpmUpdateEvent, WpmPayload);
+impl_payload_wrapper!(WpmUpdateEvent, u16);
 
 /// LED indicator state changed event
 #[event(channel_size = crate::LED_INDICATOR_EVENT_CHANNEL_SIZE, pubs = crate::LED_INDICATOR_EVENT_PUB_SIZE, subs = crate::LED_INDICATOR_EVENT_SUB_SIZE)]
 #[derive(Clone, Copy, Debug, PartialEq, Eq)]
 #[cfg_attr(feature = "defmt", derive(defmt::Format))]
-pub struct LedIndicatorEvent(pub LedPayload);
+pub struct LedIndicatorEvent(pub LedIndicator);
 
 impl LedIndicatorEvent {
     pub fn new(indicator: LedIndicator) -> Self {
-        Self(LedPayload { indicator })
+        Self(indicator)
     }
 }
 
-impl_payload_wrapper!(LedIndicatorEvent, LedPayload);
+impl_payload_wrapper!(LedIndicatorEvent, LedIndicator);
 
 /// Sleep state changed event
 #[event(channel_size = crate::SLEEP_STATE_EVENT_CHANNEL_SIZE, pubs = crate::SLEEP_STATE_EVENT_PUB_SIZE, subs = crate::SLEEP_STATE_EVENT_SUB_SIZE)]
 #[derive(Clone, Copy, Debug, PartialEq, Eq)]
 #[cfg_attr(feature = "defmt", derive(defmt::Format))]
-pub struct SleepStateEvent(pub SleepPayload);
+pub struct SleepStateEvent(pub bool);
 
 impl SleepStateEvent {
     pub fn new(sleeping: bool) -> Self {
-        Self(SleepPayload { sleeping })
+        Self(sleeping)
     }
 }
 
-impl_payload_wrapper!(SleepStateEvent, SleepPayload);
+impl_payload_wrapper!(SleepStateEvent, bool);
