+++ rmk/src/split/peripheral.rs
@@ -10,7 +10,7 @@ use {super::ble::PeerAddress, crate::channel::FLASH_CHANNEL};
 use {
     crate::event::{BatteryStatusEvent, ChargingStateEvent, EventSubscriber},
     crate::storage::Storage,
-    rmk_types::event::BatteryStatus,
+    rmk_types::battery::BatteryStatus,
     embedded_storage_async::nor_flash::NorFlash,
     trouble_host::prelude::*,
 };
